requirements.txt
# Endpoint Security Agent - Dependencies
# Core
watchdog>=3.0.0
boto3>=1.34.0
pyyaml>=6.0
psutil>=5.9.0

# Windows-specific
pywin32>=306
wmi>=1.5.1

# Network monitoring
scapy>=2.5.0
# pydivert>=2.1.0  # Requires WinDivert driver

# Encryption
cryptography>=41.0.0

# API/Dashboard Backend
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.1.0

# Utilities
aiofiles>=23.2.0
httpx>=0.26.0
schedule>=1.2.0
python-dateutil>=2.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
mypy>=1.8.0
